fix: multiarch image builds by using host architecture for jq and crane  (#869)
This PR builds on #794 to fix the failing CI jobs. 

Multiarch image builds can fail due to the target architecture for jq
and crane. For example if you are building an arm64 image on a x86
machine a failure similar to this will occur:

```
line 52: ../jq_linux_arm64/jq: cannot execute binary file: Exec format error
```
As jq and crane run on the host building the images the host
architecture should be used rather than the target arch.

---------

Signed-off-by: Brian Carey <bcarey@redhat.com>
Co-authored-by: Lionel Fleury <lionel_fleury@hotmail.com>